+++ library/src/main/java/com/heartlandpaymentsystems/library/entities/SurchargeCalculation.java
@@ -1,0 +1,9 @@
+package com.heartlandpaymentsystems.library.entities;
+
+
+public class SurchargeCalculation {
+
+    public SurchargeCalculation(boolean isPreTax){
+
+    }
+}