+++ library/src/main/java/com/heartlandpaymentsystems/library/terminals/c2x/C2XDevice.java
@@ -34,6 +34,7 @@ import com.tsys.payments.library.domain.TerminalInfo;
 import com.tsys.payments.library.domain.TransactionConfiguration;
 import com.tsys.payments.library.domain.TransactionRequest;
 import com.tsys.payments.library.domain.TransactionResponse;
+import com.tsys.payments.library.enums.CardholderInteractionType;
 import com.tsys.payments.library.enums.ConnectionType;
 import com.tsys.payments.library.enums.CurrencyCode;
 import com.tsys.payments.library.enums.TerminalAuthenticationCapability;
@@ -283,6 +284,10 @@ public class C2XDevice implements IDevice {
             timberPlanted = true;
         }
 
+        if(connectionConfig.isSurchargeEnabled()){
+            LibraryConfigHelper.setSurchargeEnabled(connectionConfig.isSurchargeEnabled());
+        }
+
         transactionConfig = new TransactionConfiguration();
         transactionConfig.setChipEnabled(true);
         transactionConfig.setQuickChipEnabled(true);
@@ -410,6 +415,7 @@ public class C2XDevice implements IDevice {
         cr.setCardholderInteractionType(info.getCardholderInteractionType());
         cr.setCommercialCardDataFields(info.getCommercialCardDataFields());
         cr.setFinalTransactionAmount(info.getFinalTransactionAmount());
+        cr.setSurchargeAmount(info.getFinalSurchargeAmount());
         cr.setSupportedApplications(info.getSupportedApplications());
         return cr;
     }
@@ -529,7 +535,46 @@ public class C2XDevice implements IDevice {
         @Override
         public void onCardholderInteractionRequested(CardholderInteractionRequest cardholderInteractionRequest) {
             if (transactionListener != null) {
-                transactionListener.onCardholderInteractionRequested(map(cardholderInteractionRequest));
+                if(cardholderInteractionRequest.getCardholderInteractionType() ==
+                        CardholderInteractionType.SURCHARGE_REQUESTED){
+                    Long amountBefore = cardholderInteractionRequest.getFinalTransactionAmount();
+                    float surcharge = amountBefore * 0.03f;
+                    Long finalAmount = (long)(amountBefore + surcharge);
+                    cardholderInteractionRequest.setSurchargeAmount((long)surcharge);
+                    cardholderInteractionRequest.setFinalTransactionAmount(finalAmount);
+                }
+                boolean interactionHandled = transactionListener.onCardholderInteractionRequested(map(cardholderInteractionRequest));
+                if (!interactionHandled) {
+                    CardholderInteractionResult result;
+                    switch (cardholderInteractionRequest.getCardholderInteractionType()) {
+                        case EMV_APPLICATION_SELECTION:
+                            String[] applications =
+                                    cardholderInteractionRequest.getSupportedApplications();
+                            // send result
+                            result = new CardholderInteractionResult(
+                                    cardholderInteractionRequest.getCardholderInteractionType()
+                            );
+                            result.setSelectedAidIndex(0);
+                            sendCardholderInteractionResult(result);
+                            break;
+                        case SURCHARGE_REQUESTED:
+                            result = new CardholderInteractionResult(
+                                    CardholderInteractionType.CARDHOLDER_SURCHARGE_CONFIRMATION);
+                            result.setFinalAmountConfirmed(false);
+                            sendCardholderInteractionResult(result);
+                            Timber.e("Surcharge confirmation was not handled by client application, cancelling transaction");
+                            break;
+                        case FINAL_AMOUNT_CONFIRMATION:
+                            result = new CardholderInteractionResult(
+                                    cardholderInteractionRequest.getCardholderInteractionType()
+                            );
+                            result.setFinalAmountConfirmed(true);
+                            sendCardholderInteractionResult(result);
+                            break;
+                        default:
+                            break;
+                    }
+                }
             }
         }